adding more implementation
adding controller implementation, model web for update request, notes, and broken dependencies, etc

+++ model/web/category_update_request.go
@@ -1,6 +1,6 @@
 package web
 
 type CategoryUpdateRequest struct {
-	Id           int64  `validate:"required"`
+	Id           int    `validate:"required"`
 	Namakategori string `validate:"required,max=255,min=1"`
 }
