+++ service/category_service.go
@@ -7,7 +7,7 @@ import (
 
 type CategoryService interface {
 	//Findbyid category spesifik ke id korelasi dengan API GetById
-	FindByIdGet(ctx context.Context, idKategori int64) web.CategoryResponse
+	FindByIdGet(ctx context.Context, idKategori int) web.CategoryResponse
 	//FindAll category tidak spesifik, korelasi dengan API Get
 	FindAllGet(ctx context.Context) []web.CategoryResponse
 	//Create category korelasi dengan API Post
@@ -15,5 +15,5 @@ type CategoryService interface {
 	//Update/replace category korelasi dengan API Put
 	UpdatePut(ctx context.Context, request web.CategoryUpdateRequest) web.CategoryResponse
 	//Delete category korelasi dengan API Get
-	DeleteDelete(ctx context.Context, idKategori int64)
+	DeleteDelete(ctx context.Context, idKategori int)
 }
